# swagger.yaml
openapi: 3.0.3
info:
  title: Backup Tools API
  description: |
    A comprehensive backup and synchronization API that allows users to backup data from various cloud services 
    (Google Drive, Google Photos, Gmail, Google Cloud Storage, Dropbox, Office 365, AWS S3, GitHub, Shopify, QuickBooks) 
    to a satellite storage system and vice versa.
    
    ## Features
    - **Multi-Platform Support**: Integrate with major cloud providers
    - **Automatic Backup**: Schedule and manage automated backup jobs
    - **Satellite Storage**: Store and retrieve data from satellite storage
    - **Real-time Sync**: Live synchronization capabilities
    - **Authentication**: JWT-based authentication with OAuth2 flows
    
    ## Authentication
    Most endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    Some endpoints also require an ACCESS_TOKEN header for satellite operations:
    ```
    ACCESS_TOKEN: <your-satellite-access-token>
    ```
  version: 1.0.0
  contact:
    name: Backup Tools Support
    email: support@backuptools.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8005
    description: Development server
  - url: https://backup.storx.io
    description: Production server
...
paths:
...
  /satellite-auth:
    post:
      tags:
        - Authentication
      summary: Authenticate with Satellite
      description: Authenticate user with satellite storage system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Satellite authentication token
              required:
                - token
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /google-auth:
    post:
      tags:
        - Authentication
      summary: Google OAuth Authentication
      description: Authenticate with Google services using OAuth2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: OAuth2 authorization code
                refresh_token:
                  type: string
                  description: OAuth2 refresh token
              required:
                - code
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Authentication
      summary: Google OAuth Authentication (GET)
      description: Alternative Google OAuth authentication endpoint
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Microsoft Autosync (product twin of Google autosync; filter by outlook* methods)
...
  /microsoft/users-groups:
    get:
      tags:
        - Microsoft Autosync
      summary: List Microsoft Users & Groups mailboxes
      description: Same shape as GET /users-groups but Microsoft methods only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Mailbox entities
  /microsoft/query-messages:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Outlook messages
      description: Requires REFRESH_TOKEN header. Prefer this over legacy /office365 for new Microsoft UI.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message list
  /microsoft/contacts/list:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Outlook contacts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contacts list
  /microsoft/calendar/list:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Outlook calendars
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Calendars list
  /microsoft/calendar/events/{calendarId}:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Outlook calendar events
      security:
        - bearerAuth: []
      parameters:
        - name: calendarId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Events list
  /microsoft/onedrive/flat-files:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse OneDrive flat non-folder files
      description: |
        Twin of GET /google/drive-flat-files for Microsoft OneDrive.
        Requires REFRESH_TOKEN header. Optional email/mailbox selects /users/{mailbox}/drive vs /me/drive.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Flat file list
  /microsoft/outlook/flat-files:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Outlook inbox messages
      description: |
        Twin of Gmail browse for Microsoft Outlook mail.
        Requires REFRESH_TOKEN header. Optional mailbox/email selects /users/{mailbox} vs /me.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: mailbox
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Outlook message list
  /microsoft/sharepoint/sites:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse SharePoint sites (site picker)
      description: |
        Lists all SharePoint sites in the caller's tenant (getAllSites, falling back to search).
        REFRESH_TOKEN identifies the tenant; listing uses the platform app-only token and requires
        granted admin consent plus the sharepoint_backup capability.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: SharePoint sites list
  /microsoft/sharepoint/flat-files:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse SharePoint document library flat non-folder files
      description: |
        Twin of GET /microsoft/onedrive/flat-files for a SharePoint document library drive.
        Requires REFRESH_TOKEN header and drive_id query (resolved at onboarding via GET /sites/{site_id}/drive).
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: drive_id
          in: query
          required: true
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Flat file list
  /microsoft/teams/list:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Teams (team picker)
      description: |
        Lists all Teams in the caller's tenant via the platform app-only token.
        Requires granted admin consent plus the teams_backup capability.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Teams list
  /microsoft/teams/channels:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse channels for a Team
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: team_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team channels list
  /microsoft/teams/flat-messages:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse Teams channel messages
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: team_id
          in: query
          required: true
          schema:
            type: string
        - name: channel_id
          in: query
          required: true
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Channel messages list
  /microsoft/groups/list:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse M365 Groups (group picker)
      description: |
        Lists all Microsoft 365 groups in the caller's tenant via the platform app-only token.
        Requires granted admin consent plus the groups_backup capability.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Groups list
  /microsoft/groups/flat-conversations:
    get:
      tags:
        - Microsoft Autosync
      summary: Browse M365 Group conversation threads
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          required: true
          schema:
            type: string
        - name: group_id
          in: query
          required: true
          schema:
            type: string
        - name: top
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Group conversations list
  /microsoft/outlook/corporate/domain-users:
    get:
      tags:
        - Microsoft Autosync
      summary: Detect Microsoft account type (Google twin)
      description: |
        Returns the workspace contract (account_type, tenant_id, consent, capabilities, directory).
        Read-only: never writes tenant consent or capabilities. account_type and is_admin are labels;
        admin_workspace means consent is granted and list_users is available. In that case entities[]
        holds a first page of users from the synced directory.
        Requires REFRESH_TOKEN header (Satellite convention).
      security:
        - bearerAuth: []
      parameters:
        - name: top
          in: query
          schema:
            type: integer
          description: Max directory users for admin_workspace (default 200)
      responses:
        '200':
          description: Account detection result
  /microsoft/account/detect:
    get:
      tags:
        - Microsoft Autosync
      summary: Alias of GET /microsoft/outlook/corporate/domain-users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account detection result
  /microsoft/directory/users:
    get:
      tags:
        - Microsoft Autosync
      summary: Paginated Microsoft tenant directory users (admin picker)
      description: |
        Lists users live from Microsoft Graph for the REFRESH_TOKEN's tenant (nothing is stored).
        Same response as GET /microsoft/tenants/{tid}/directory/users.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MicrosoftDirectoryPage'
        - $ref: '#/components/parameters/MicrosoftDirectoryPageSize'
        - $ref: '#/components/parameters/MicrosoftDirectorySearch'
        - $ref: '#/components/parameters/MicrosoftDirectoryDepartment'
        - $ref: '#/components/parameters/MicrosoftDirectoryOrgUnit'
        - $ref: '#/components/parameters/MicrosoftDirectoryEnabledOnly'
      responses:
        '200':
          description: Paginated directory users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicrosoftDirectoryUsersResponse'
        '403':
          description: Personal account or tenant consent not granted
  /microsoft/workspace:
    get:
      tags:
        - Microsoft Autosync
      summary: Microsoft workspace contract
      description: |
        Returns the shared Satellite <-> Backup-Tools workspace contract. With a REFRESH_TOKEN header,
        the signed-in account is resolved live. Otherwise the caller's stored Microsoft credential is used
        (optionally narrowed by tenant_id, email or project_id). Read-only.
      security:
        - bearerAuth: []
      parameters:
        - name: REFRESH_TOKEN
          in: header
          schema:
            type: string
        - name: tenant_id
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Workspace contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicrosoftWorkspaceContract'
        '404':
          description: No Microsoft credential found for this user
  /microsoft/accounts/tenants:
    get:
      tags:
        - Microsoft Tenants
      summary: List the tenants a Microsoft sign-in can reach
      description: |
        Discovers every tenant of the sign-in (home, guest, personal), checks a tenant-scoped token and
        the user's roles in each and returns their access states. The home link always exists; a
        discovery failure keeps known links and reports discovery_status. Creates the credential on first call.
      parameters:
        - $ref: '#/components/parameters/tokenKeyHeader'
        - $ref: '#/components/parameters/MicrosoftAccountIDHeader'
        - $ref: '#/components/parameters/MicrosoftHomeTenantIDHeader'
        - $ref: '#/components/parameters/MicrosoftRefreshTokenHeader'
        - in: query
          name: project_id
          schema:
            type: string
      responses:
        '200':
          description: Account and its tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicrosoftAccountTenants'
        '403':
          description: credential_forbidden or signin_required
        '404':
          description: credential_not_found
...
  /users-groups/domains:
    get:
      tags:
        - Auto Sync
      summary: List Connected Domains
      description: |
        Returns unique domain strings from connected credentials for the Users & Groups domain dropdown.
        Scoped to credentials linked to the user's non-placeholder jobs.
      parameters:
        - $ref: '#/components/parameters/tokenKeyHeader'
      responses:
        '200':
          description: Unique domain strings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersGroupsDomainsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /google/drive-root-file-names:
    get:
      tags:
        - Google Drive
      summary: List Root Files
      description: List all files in Google Drive root directory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of root files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriveFile'
...
components:
  parameters:
...
    MicrosoftDirectoryPage:
      name: page
      in: query
      schema:
        type: integer
        default: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...
security:
  - bearerAuth: []